feat:#8 ticketSaleBuffer, ticketArtistBuffer 추가 api
