+++ newket-api/src/main/kotlin/com/newket/api/controller/ticket/TicketApi.kt
@@ -9,6 +9,5 @@ object TicketApi {
         const val ON_SALE = "${BASE_URL}/on-sale"
         const val SEARCH = "${BASE_URL}/search"
         const val AUTOCOMPLETE = "${BASE_URL}/autocomplete"
-        const val FETCH = "${BASE_URL}/fetch"
     }
 }
