+++ newket-application/src/main/kotlin/com/newket/application/admin/dto/AddTicketArtistsRequest.kt
@@ -1,0 +1,8 @@
+package com.newket.application.admin.dto
+
+import com.newket.client.crawling.CreateTicketRequest
+
+data class AddTicketArtistsRequest (
+    val ticketId: Long,
+    val artists: List<CreateTicketRequest.Artist>
+)
