+++ newket-application/src/main/kotlin/com/newket/application/admin/dto/TextDto.kt
@@ -1,0 +1,5 @@
+package com.newket.application.admin.dto
+
+data class TextDto (
+    val text: String
+)
