--- newket-application/src/main/kotlin/com/newket/application/ticket/dto/TicketUrlRequest.kt
@@ -1,5 +1,0 @@
-package com.newket.application.ticket.dto
-
-data class TicketUrlRequest (
-    val url: String
-)
